#Cyto/requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
eventlet>=0.33.0
python-socketio>=5.9.0
